Add leaf_fanouts fixture and eos_to_linux_intf helper function
