+++ tests/qos/qos_fixtures.py
@@ -17,5 +17,20 @@ def conn_graph_facts(testbed_devices):
 	
     return result
 
+@pytest.fixture(scope = "module")
+def leaf_fanouts(conn_graph_facts):
+    """
+    @summary: Fixture for getting the list of leaf fanout switches
+    @param conn_graph_facts: Topology connectivity information
+    @return: Return the list of leaf fanout switches
+    """
+    leaf_fanouts = []
+    conn_facts = conn_graph_facts['device_conn']
+    
+    """ for each interface of DUT """
+    for intf in conn_facts:
+        peer_device = conn_facts[intf]['peerdevice']
+        if peer_device not in leaf_fanouts:
+            leaf_fanouts.append(peer_device)
 
-
+    return leaf_fanouts
