+++ tests/qos/qos_helpers.py
@@ -1,0 +1,7 @@
+def eos_to_linux_intf(eos_intf_name):
+    """
+    @Summary: Map EOS's interface name to Linux's interface name
+    @param eos_intf_name: Interface name in EOS
+    @return: Return the interface name in Linux
+    """
+    return eos_intf_name.replace('Ethernet', 'et').replace('/', '_')
