+++ tests/qos/test_pfc_counters.py
@@ -1,5 +1,6 @@
 from ansible_host import AnsibleHost
-from qos_fixtures import conn_graph_facts
+from qos_fixtures import conn_graph_facts, leaf_fanouts
+from qos_helpers import eos_to_linux_intf
 import os
 import time
 
@@ -20,44 +21,36 @@ PKT_COUNT = 10
 """ Number of switch priorities """ 
 PRIO_COUNT = 8
 
-def setup_testbed(ansible_adhoc, testbed, conn_graph_facts):
+def setup_testbed(ansible_adhoc, testbed, leaf_fanouts):
     """
     @Summary: Set up the testbed, including clearing counters, and copying the PFC generator to the leaf fanout switches.
     @param ansible_adhoc: Fixture provided by the pytest-ansible package. Source of the various device objects. It is
     mandatory argument for the class constructors.
     @param testbed: Testbed information
-    @param conn_graph_facts: Testbed topology connectivity information
+    @param leaf_fanouts: Leaf fanout switches
     """
     dut_hostname = testbed['dut']
     dut_ans = AnsibleHost(ansible_adhoc, dut_hostname)
     """ Clear PFC counters """
     dut_ans.sonic_pfc_counters(method = "clear")
 
-    conn_facts = conn_graph_facts['device_conn']	
-    """ Get all the leaf fanout switches """
-    leaf_fanouts = []
-    """ For each interface of DUT"""
-    for intf in conn_facts:
-        peer_device = conn_facts[intf]['peerdevice']
-        if peer_device not in leaf_fanouts:
-            leaf_fanouts.append(peer_device)
-	
     """ Copy the PFC generator to all the leaf fanout switches """
     for peer_device in leaf_fanouts:
         peerdev_ans = AnsibleHost(ansible_adhoc, peer_device)
         file_src = os.path.join(os.path.dirname(__file__), PFC_GEN_FILE_RELATIVE_PATH)
         peerdev_ans.copy(src = file_src, dest = PFC_GEN_FILE_DEST, force = True)
 
-def run_pfc(ansible_adhoc, testbed, conn_graph_facts, pause_time):
+def run_pfc(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts, pause_time):
     """
     @Summary: Run the priority-based flow control (PFC) test case
     @param ansible_adhoc: Fixture provided by the pytest-ansible package. Source of the various device objects. It is
     mandatory argument for the class constructors.
     @param testbed: Testbed information
     @param conn_graph_facts: Testbed topology connectivity information
+    @param leaf_fanouts: Leaf fanout switches
     @param pause_time: Pause time quanta (0-65535) in the frame. 0 means unpause.
     """
-    setup_testbed(ansible_adhoc, testbed, conn_graph_facts)
+    setup_testbed(ansible_adhoc, testbed, leaf_fanouts)
     conn_facts = conn_graph_facts['device_conn']
      
     """ Generate PFC packets for all the priority queues of all the interfaces """
@@ -84,16 +77,17 @@ def run_pfc(ansible_adhoc, testbed, conn_graph_facts, pause_time):
         assert 'Rx' in counter_facts[intf]
         assert counter_facts[intf]['Rx'] == [str(PKT_COUNT)] * PRIO_COUNT
 
-def run_fc(ansible_adhoc, testbed, conn_graph_facts, pause_time):
+def run_fc(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts, pause_time):
     """
     @Summary: Run the flow control (FC) test case
     @param ansible_adhoc: Fixture provided by the pytest-ansible package. Source of the various device objects. It is
     mandatory argument for the class constructors.
     @param testbed: Testbed information
     @param conn_graph_facts: Testbed topology connectivity information
+    @param leaf_fanouts: Leaf fanout switches
     @param pause_time: Pause time quanta (0-65535) in the frame. 0 means unpause.
     """
-    setup_testbed(ansible_adhoc, testbed, conn_graph_facts)
+    setup_testbed(ansible_adhoc, testbed, leaf_fanouts)
     conn_facts = conn_graph_facts['device_conn']
 
     """ Generate flow control packets for all the interfaces """
@@ -119,26 +113,18 @@ def run_fc(ansible_adhoc, testbed, conn_graph_facts, pause_time):
         assert 'Rx' in counter_facts[intf]
         assert counter_facts[intf]['Rx'] == ['0'] * PRIO_COUNT
 
-def eos_to_linux_intf(eos_intf_name):
-    """
-    @Summary: Map EOS's interface name to Linux's interface name
-    @param eos_intf_name: Interface name in EOS
-    @return: Return the interface name in Linux 
-    """
-    return eos_intf_name.replace('Ethernet', 'et').replace('/', '_')
-
-def test_pfc_pause(ansible_adhoc, testbed, conn_graph_facts):
+def test_pfc_pause(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts):
     """ @Summary: Run PFC pause frame (pause time quanta > 0) tests """
-    run_pfc(ansible_adhoc, testbed, conn_graph_facts, 65535)
+    run_pfc(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts, 65535)
 			
-def test_pfc_unpause(ansible_adhoc, testbed, conn_graph_facts):
+def test_pfc_unpause(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts):
     """ @Summary: Run PFC unpause frame (pause time quanta = 0) tests """
-    run_pfc(ansible_adhoc, testbed, conn_graph_facts, 0)        
+    run_pfc(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts, 0)        
 
-def test_fc_pause(ansible_adhoc, testbed, conn_graph_facts):
+def test_fc_pause(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts):
     """ @Summary: Run FC pause frame (pause time quanta > 0) tests """
-    run_fc(ansible_adhoc, testbed, conn_graph_facts, 65535)
+    run_fc(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts, 65535)
 
-def test_fc_unpause(ansible_adhoc, testbed, conn_graph_facts):
+def test_fc_unpause(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts):
     """ @Summary: Run FC pause frame (pause time quanta = 0) tests """ 
-    run_fc(ansible_adhoc, testbed, conn_graph_facts, 0) 	
+    run_fc(ansible_adhoc, testbed, conn_graph_facts, leaf_fanouts, 0)
